@@ -146,7 +146,8 @@ describe("loop.contacts", function() {
     it("should show the gravatars promo box", function() {
       listView = TestUtils.renderIntoDocument(
         React.createElement(loop.contacts.ContactsList, {
-          notifications: notifications
+          notifications: notifications,
+          startForm: function() {}
         }));
 
       var promo = listView.getDOMNode().querySelector(".contacts-gravatar-promo");
@@ -166,7 +167,8 @@ describe("loop.contacts", function() {
       };
       listView = TestUtils.renderIntoDocument(
         React.createElement(loop.contacts.ContactsList, {
-          notifications: notifications
+          notifications: notifications,
+          startForm: function() {}
         }));
 
       var promo = listView.getDOMNode().querySelector(".contacts-gravatar-promo");
@@ -178,7 +180,8 @@ describe("loop.contacts", function() {
     it("should hide the gravatars promo box when the 'use' button is clicked", function() {
       listView = TestUtils.renderIntoDocument(
         React.createElement(loop.contacts.ContactsList, {
-          notifications: notifications
+          notifications: notifications,
+          startForm: function() {}
         }));
 
       React.addons.TestUtils.Simulate.click(listView.getDOMNode().querySelector(
@@ -193,7 +196,8 @@ describe("loop.contacts", function() {
     it("should should set the prefs correctly when the 'use' button is clicked", function() {
       listView = TestUtils.renderIntoDocument(
         React.createElement(loop.contacts.ContactsList, {
-          notifications: notifications
+          notifications: notifications,
+          startForm: function() {}
         }));
 
       React.addons.TestUtils.Simulate.click(listView.getDOMNode().querySelector(
@@ -207,7 +211,8 @@ describe("loop.contacts", function() {
     it("should hide the gravatars promo box when the 'close' button is clicked", function() {
       listView = TestUtils.renderIntoDocument(
         React.createElement(loop.contacts.ContactsList, {
-          notifications: notifications
+          notifications: notifications,
+          startForm: function() {}
         }));
 
       React.addons.TestUtils.Simulate.click(listView.getDOMNode().querySelector(
@@ -220,7 +225,8 @@ describe("loop.contacts", function() {
     it("should set prefs correctly when the 'close' button is clicked", function() {
       listView = TestUtils.renderIntoDocument(
         React.createElement(loop.contacts.ContactsList, {
-          notifications: notifications
+          notifications: notifications,
+          startForm: function() {}
         }));
 
       React.addons.TestUtils.Simulate.click(listView.getDOMNode().querySelector(
@@ -242,7 +248,8 @@ describe("loop.contacts", function() {
 
       listView = TestUtils.renderIntoDocument(
         React.createElement(loop.contacts.ContactsList, {
-          notifications: notifications
+          notifications: notifications,
+          startForm: function() {}
         }));
     });
 
@@ -299,8 +306,10 @@ describe("loop.contacts", function() {
 
         beforeEach(function() {
           view = TestUtils.renderIntoDocument(
-            React.createElement(
-              loop.contacts.ContactDetailsForm, {mode: "add"}));
+            React.createElement(loop.contacts.ContactDetailsForm, {
+              mode: "add",
+              selectTab: function() {}
+            }));
         });
 
         it("should render 'add' header", function() {
@@ -409,8 +418,10 @@ describe("loop.contacts", function() {
 
         beforeEach(function() {
           view = TestUtils.renderIntoDocument(
-            React.createElement(
-              loop.contacts.ContactDetailsForm, {mode: "edit"}));
+            React.createElement(loop.contacts.ContactDetailsForm, {
+              mode: "edit",
+              selectTab: function() {}
+            }));
         });
 
         it("should render 'edit' header", function() {
