@@ -778,7 +778,7 @@ describe("loop.panel", function() {
        "conversation button",
       function() {
         navigator.mozLoop.userProfile = {email: fakeEmail};
-        var view = createTestComponent();
+        var view = createTestComponent(false);
 
         TestUtils.Simulate.click(view.getDOMNode().querySelector(".new-room-button"));
 
@@ -801,7 +801,7 @@ describe("loop.panel", function() {
         });
       };
 
-      var view = createTestComponent();
+      var view = createTestComponent(false);
 
       // Simulate being visible
       view.onDocumentVisible();
@@ -841,7 +841,7 @@ describe("loop.panel", function() {
         });
       };
 
-      var view = createTestComponent();
+      var view = createTestComponent(false);
 
       // Simulate being visible
       view.onDocumentVisible();
@@ -859,7 +859,7 @@ describe("loop.panel", function() {
         });
       };
 
-      var view = createTestComponent();
+      var view = createTestComponent(false);
 
       view.setState({ checked: true });
 
@@ -878,7 +878,7 @@ describe("loop.panel", function() {
         });
       };
 
-      var view = createTestComponent();
+      var view = createTestComponent(false);
 
       // Simulate being visible
       view.onDocumentVisible();
@@ -896,7 +896,7 @@ describe("loop.panel", function() {
         });
       };
 
-      var view = createTestComponent();
+      var view = createTestComponent(false);
 
       view.onDocumentVisible();
 
@@ -913,7 +913,7 @@ describe("loop.panel", function() {
         });
       };
 
-      var view = createTestComponent();
+      var view = createTestComponent(false);
 
       // Simulate being visible
       view.onDocumentVisible();
@@ -933,7 +933,7 @@ describe("loop.panel", function() {
         });
       };
 
-      var view = createTestComponent();
+      var view = createTestComponent(false);
 
       // Simulate being visible.
       view.onDocumentVisible();
