@@ -256,6 +256,9 @@ describe("loop.shared.views", function() {
     var hangup, publishStream;
 
     function mountTestComponent(props) {
+      props = _.extend({
+        dispatcher: dispatcher
+      }, props || {});
       return TestUtils.renderIntoDocument(
         React.createElement(sharedViews.ConversationToolbar, props));
     }
@@ -362,6 +365,9 @@ describe("loop.shared.views", function() {
     var fakeSDK, fakeSessionData, fakeSession, fakePublisher, model, fakeAudio;
 
     function mountTestComponent(props) {
+      props = _.extend({
+        dispatcher: dispatcher
+      }, props || {});
       return TestUtils.renderIntoDocument(
         React.createElement(sharedViews.ConversationView, props));
     }
@@ -676,6 +682,9 @@ describe("loop.shared.views", function() {
     var coll, view, testNotif;
 
     function mountTestComponent(props) {
+      props = _.extend({
+        key: 0
+      }, props || {});
       return TestUtils.renderIntoDocument(
         React.createElement(sharedViews.NotificationListView, props));
     }
@@ -831,7 +840,11 @@ describe("loop.shared.views", function() {
 
     function mountTestComponent(extraProps) {
       var props = _.extend({
-        dispatcher: dispatcher
+        allowClick: false,
+        description: "test",
+        dispatcher: dispatcher,
+        showContextTitle: false,
+        useDesktopPaths: false
       }, extraProps);
       return TestUtils.renderIntoDocument(
         React.createElement(sharedViews.ContextUrlView, props));
@@ -913,6 +926,9 @@ describe("loop.shared.views", function() {
     var view;
 
     function mountTestComponent(props) {
+      props = _.extend({
+        isLoading: false
+      }, props || {});
       return TestUtils.renderIntoDocument(
         React.createElement(sharedViews.MediaView, props));
     }
