@@ -4,9 +4,39 @@
 
 #include "nsAutoCompleteSimpleResult.h"
 
+#define CHECK_MATCH_INDEX(_index, _insert)                                     \
+  if (_index < 0 ||                                                            \
+      static_cast<MatchesArray::size_type>(_index) > mMatches.Length() ||      \
+      (!_insert && static_cast<MatchesArray::size_type>(_index) == mMatches.Length())) { \
+    MOZ_ASSERT(false, "Trying to use an invalid index on mMatches");           \
+    return NS_ERROR_ILLEGAL_VALUE;                                             \
+  }                                                                            \
+
 NS_IMPL_ISUPPORTS(nsAutoCompleteSimpleResult,
                   nsIAutoCompleteResult,
                   nsIAutoCompleteSimpleResult)
+
+struct AutoCompleteSimpleResultMatch
+{
+  AutoCompleteSimpleResultMatch(const nsAString& aValue,
+                                const nsAString& aComment,
+                                const nsAString& aImage,
+                                const nsAString& aStyle,
+                                const nsAString& aFinalCompleteValue)
+    : mValue(aValue)
+    , mComment(aComment)
+    , mImage(aImage)
+    , mStyle(aStyle)
+    , mFinalCompleteValue(aFinalCompleteValue)
+  {
+  }
+
+  nsString mValue;
+  nsString mComment;
+  nsString mImage;
+  nsString mStyle;
+  nsString mFinalCompleteValue;
+};
 
 nsAutoCompleteSimpleResult::nsAutoCompleteSimpleResult() :
   mDefaultIndex(-1),
@@ -87,58 +117,47 @@ nsAutoCompleteSimpleResult::SetTypeAheadResult(bool aTypeAheadResult)
 }
 
 NS_IMETHODIMP
+nsAutoCompleteSimpleResult::InsertMatchAt(int32_t aIndex,
+                                          const nsAString& aValue,
+                                          const nsAString& aComment,
+                                          const nsAString& aImage,
+                                          const nsAString& aStyle,
+                                          const nsAString& aFinalCompleteValue)
+{
+  CHECK_MATCH_INDEX(aIndex, true);
+
+  AutoCompleteSimpleResultMatch match(aValue, aComment, aImage, aStyle, aFinalCompleteValue);
+
+  if (!mMatches.InsertElementAt(aIndex, match)) {
+    return NS_ERROR_OUT_OF_MEMORY;
+  }
+
+  return NS_OK;
+}
+
+NS_IMETHODIMP
 nsAutoCompleteSimpleResult::AppendMatch(const nsAString& aValue,
                                         const nsAString& aComment,
                                         const nsAString& aImage,
                                         const nsAString& aStyle,
                                         const nsAString& aFinalCompleteValue)
 {
-  CheckInvariants();
-
-  if (! mValues.AppendElement(aValue))
-    return NS_ERROR_OUT_OF_MEMORY;
-  if (! mComments.AppendElement(aComment)) {
-    mValues.RemoveElementAt(mValues.Length() - 1);
-    return NS_ERROR_OUT_OF_MEMORY;
-  }
-  if (! mImages.AppendElement(aImage)) {
-    mValues.RemoveElementAt(mValues.Length() - 1);
-    mComments.RemoveElementAt(mComments.Length() - 1);
-    return NS_ERROR_OUT_OF_MEMORY;
-  }
-  if (! mStyles.AppendElement(aStyle)) {
-    mValues.RemoveElementAt(mValues.Length() - 1);
-    mComments.RemoveElementAt(mComments.Length() - 1);
-    mImages.RemoveElementAt(mImages.Length() - 1);
-    return NS_ERROR_OUT_OF_MEMORY;
-  }
-  if (!mFinalCompleteValues.AppendElement(aFinalCompleteValue)) {
-    mValues.RemoveElementAt(mValues.Length() - 1);
-    mComments.RemoveElementAt(mComments.Length() - 1);
-    mImages.RemoveElementAt(mImages.Length() - 1);
-    mStyles.RemoveElementAt(mStyles.Length() - 1);
-    return NS_ERROR_OUT_OF_MEMORY;
-  }
-  return NS_OK;
+  return InsertMatchAt(mMatches.Length(), aValue, aComment, aImage, aStyle,
+                       aFinalCompleteValue);
 }
 
 NS_IMETHODIMP
 nsAutoCompleteSimpleResult::GetMatchCount(uint32_t *aMatchCount)
 {
-  CheckInvariants();
-
-  *aMatchCount = mValues.Length();
+  *aMatchCount = mMatches.Length();
   return NS_OK;
 }
 
 NS_IMETHODIMP
 nsAutoCompleteSimpleResult::GetValueAt(int32_t aIndex, nsAString& _retval)
 {
-  NS_ENSURE_TRUE(aIndex >= 0 && aIndex < int32_t(mValues.Length()),
-                 NS_ERROR_ILLEGAL_VALUE);
-  CheckInvariants();
-
-  _retval = mValues[aIndex];
+  CHECK_MATCH_INDEX(aIndex, false);
+  _retval = mMatches[aIndex].mValue;
   return NS_OK;
 }
 
@@ -151,30 +170,24 @@ nsAutoCompleteSimpleResult::GetLabelAt(int32_t aIndex, nsAString& _retval)
 NS_IMETHODIMP
 nsAutoCompleteSimpleResult::GetCommentAt(int32_t aIndex, nsAString& _retval)
 {
-  NS_ENSURE_TRUE(aIndex >= 0 && aIndex < int32_t(mComments.Length()),
-                 NS_ERROR_ILLEGAL_VALUE);
-  CheckInvariants();
-  _retval = mComments[aIndex];
+  CHECK_MATCH_INDEX(aIndex, false);
+  _retval = mMatches[aIndex].mComment;
   return NS_OK;
 }
 
 NS_IMETHODIMP
 nsAutoCompleteSimpleResult::GetImageAt(int32_t aIndex, nsAString& _retval)
 {
-  NS_ENSURE_TRUE(aIndex >= 0 && aIndex < int32_t(mImages.Length()),
-                 NS_ERROR_ILLEGAL_VALUE);
-  CheckInvariants();
-  _retval = mImages[aIndex];
+  CHECK_MATCH_INDEX(aIndex, false);
+  _retval = mMatches[aIndex].mImage;
   return NS_OK;
 }
 
 NS_IMETHODIMP
 nsAutoCompleteSimpleResult::GetStyleAt(int32_t aIndex, nsAString& _retval)
 {
-  NS_ENSURE_TRUE(aIndex >= 0 && aIndex < int32_t(mStyles.Length()),
-                 NS_ERROR_ILLEGAL_VALUE);
-  CheckInvariants();
-  _retval = mStyles[aIndex];
+  CHECK_MATCH_INDEX(aIndex, false);
+  _retval = mMatches[aIndex].mStyle;
   return NS_OK;
 }
 
@@ -182,12 +195,10 @@ NS_IMETHODIMP
 nsAutoCompleteSimpleResult::GetFinalCompleteValueAt(int32_t aIndex,
                                                     nsAString& _retval)
 {
-  NS_ENSURE_TRUE(aIndex >= 0 && aIndex < int32_t(mFinalCompleteValues.Length()),
-                 NS_ERROR_ILLEGAL_VALUE);
-  CheckInvariants();
-  _retval = mFinalCompleteValues[aIndex];
+  CHECK_MATCH_INDEX(aIndex, false);
+  _retval = mMatches[aIndex].mFinalCompleteValue;
   if (_retval.Length() == 0)
-    _retval = mValues[aIndex];
+    _retval = mMatches[aIndex].mValue;
   return NS_OK;
 }
 
@@ -202,18 +213,13 @@ NS_IMETHODIMP
 nsAutoCompleteSimpleResult::RemoveValueAt(int32_t aRowIndex,
                                           bool aRemoveFromDb)
 {
-  NS_ENSURE_TRUE(aRowIndex >= 0 && aRowIndex < int32_t(mValues.Length()),
-                 NS_ERROR_ILLEGAL_VALUE);
+  CHECK_MATCH_INDEX(aRowIndex, false);
 
-  nsAutoString removedValue(mValues[aRowIndex]);
-  mValues.RemoveElementAt(aRowIndex);
-  mComments.RemoveElementAt(aRowIndex);
-  mImages.RemoveElementAt(aRowIndex);
-  mStyles.RemoveElementAt(aRowIndex);
-  mFinalCompleteValues.RemoveElementAt(aRowIndex);
+  nsString value = mMatches[aRowIndex].mValue;
+  mMatches.RemoveElementAt(aRowIndex);
 
   if (mListener)
-    mListener->OnValueRemoved(this, removedValue, aRemoveFromDb);
+    mListener->OnValueRemoved(this, value, aRemoveFromDb);
 
   return NS_OK;
 }
