@@ -17,7 +17,7 @@
 
 // This is the schema version. Update it at any schema change and add a
 // corresponding migrateVxx method below.
-#define DATABASE_SCHEMA_VERSION 29
+#define DATABASE_SCHEMA_VERSION 30
 
 // Fired after Places inited.
 #define TOPIC_PLACES_INIT_COMPLETE "places-init-complete"
@@ -255,9 +255,7 @@ protected:
    * Helpers used by schema upgrades.
    */
   nsresult MigrateV13Up();
-  nsresult MigrateV14Up();
   nsresult MigrateV15Up();
-  nsresult MigrateV16Up();
   nsresult MigrateV17Up();
   nsresult MigrateV18Up();
   nsresult MigrateV19Up();
@@ -270,7 +268,7 @@ protected:
   nsresult MigrateV26Up();
   nsresult MigrateV27Up();
   nsresult MigrateV28Up();
-  nsresult MigrateV29Up();
+  nsresult MigrateV30Up();
 
   nsresult UpdateBookmarkRootTitles();
 
